backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic-settings==2.5.0
python-multipart==0.0.12

# Supabase
supabase==2.9.1

# Vector Store
pinecone-client==5.0.1

# LangChain + Google GenAI
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-pinecone>=0.2.0

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.2

# Text Chunking
langchain-text-splitters==0.3.2

# SSE Streaming
sse-starlette>=2.0.0
